crud/pyproject.toml
[tool.poetry]
name = "crud"
version = "0.1.0"
description = ""
authors = ["dyo78 <maharjan.diyaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
